Did a lot here:
So refactored a bit. The views are purely used to display the content of whatever component it should be showing. So AddOnView has 1 job which is to display the AddOnCard, same with other components. This way we (the royal 'we') can adhere to the SOLID principles. Makes wayyyy more sense to do it this way